requirements.txt
asttokens==2.4.1
attrs==23.1.0
Automat==22.10.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==41.0.7
cssselect==1.2.0
decorator==5.1.1
executing==2.0.1
filelock==3.13.1
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
ipython==8.19.0
itemadapter==0.8.0
itemloaders==1.1.0
jedi==0.19.1
jmespath==1.0.1
lxml==4.9.4
matplotlib-inline==0.1.6
mysql-connector-python==8.2.0
packaging==23.2
parsel==1.8.1
parso==0.8.3
pexpect==4.9.0
prompt-toolkit==3.0.43
Protego==0.3.0
protobuf==4.21.12
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyDispatcher==2.0.7
Pygments==2.17.2
pyOpenSSL==23.3.0
queuelib==1.6.2
requests==2.31.0
requests-file==1.5.1
Scrapy==2.11.0
service-identity==23.1.0
six==1.16.0
stack-data==0.6.3
tldextract==5.1.1
traitlets==5.14.0
Twisted==22.10.0
typing_extensions==4.9.0
urllib3==2.1.0
w3lib==2.1.2
wcwidth==0.2.12
zope.interface==6.1
